Prepare exception for restaurants.

+++ Foodie.Meals.Application/Functions/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Foodie.Meals.Application.Contracts.Infrastructure.Repositories;
+using Foodie.Meals.Domain.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -21,6 +22,10 @@ namespace Foodie.Meals.Application.Functions.Restaurants.Commands.DeleteRestaura
         public async Task<Unit> Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
             var restaurantToDelete = await restaurantRepository.GetByIdAsync(request.RestaurantId);
+
+            if (restaurantToDelete == null)
+                throw new RestaurantNotFoundException(request.RestaurantId);
+
             await restaurantRepository.DeleteAsync(restaurantToDelete);
             return new Unit();
         }
